requirements.txt
discord.py
python-dotenv
pymongo
certifi